refactor: extract history events table buttons
